MKMPT-19: Code review. Made helper functions private.
