+++ iOSKMP/iOSKMP/Scenes/Home/ViewModel/CurrencyExchangeViewModel.swift
@@ -174,3 +174,30 @@ import CurrencyExchangeKMP
         alertError = AlertError.localDataFetchFailed
     }
 }
+
+// MARK: - Helpers
+
+private extension Collection where Element == CurrencyExchangeRate {
+    func mapToCurrencyExchangeRateViewModels() -> [CurrencyExchangeRateViewModel] {
+        map {
+            CurrencyExchangeRateViewModel(
+                id: $0.id,
+                code: $0.code,
+                currencyName: $0.currencyName,
+                exchangeRate: $0.exchangeRate
+            )
+        }
+    }
+}
+
+private extension Collection where Element == CurrencyExchange {
+    func mapToCurrencyExchangeListViewModels() -> [CurrencyExchangeListViewModel] {
+        map {
+            CurrencyExchangeListViewModel(
+                id: $0.id,
+                date: $0.date,
+                exchangeRates: $0.exchangeRates.mapToCurrencyExchangeRateViewModels()
+            )
+        }
+    }
+}
